pyproject.toml

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "learn-ai"
version = "0.1.0"
description = "AI tutor with memory and RAG"
readme = "README.md"
requires-python = ">=3.10"

authors = [
  { name = "Your Name", email = "you@example.com" }
]

dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
]

[project.scripts]
learn-ai = "src.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
